+++ src/versioningit/git.py
@@ -1,7 +1,8 @@
+from dataclasses import dataclass
 from pathlib import Path
 import re
 import subprocess
-from typing import Any, Optional, Union
+from typing import Any, List, Optional, Union
 from .core import VCSDescription
 from .errors import NoTagError, NotVCSError
 from .logging import log, warn_extra_fields
@@ -13,6 +14,51 @@ from .util import (
     runcmd,
     str_guard,
 )
+
+DEFAULT_DATE = fromtimestamp(0)
+
+
+@dataclass
+class GitRepo:
+    path: Union[str, Path]
+
+    def ensure_is_repo(self) -> None:
+        try:
+            runcmd(
+                "git",
+                "-C",
+                self.path,
+                "rev-parse",
+                "--git-dir",
+                stdout=subprocess.DEVNULL,
+                stderr=subprocess.DEVNULL,
+            )
+        except FileNotFoundError:
+            raise NotVCSError("Git not installed; assuming this isn't a Git repository")
+        except subprocess.CalledProcessError:
+            raise NotVCSError(f"{self.path} is not a Git repository")
+
+    def read(self, *args: str, **kwargs: Any) -> str:
+        return readcmd("git", "-C", self.path, *args, **kwargs)
+
+    def describe(self, match: List[str], exclude: List[str]) -> str:
+        cmd = ["describe", "--tags", "--long", "--dirty", "--always"]
+        for pat in match:
+            cmd.append(f"--match={pat}")
+        for pat in exclude:
+            cmd.append(f"--exclude={pat}")
+        try:
+            return self.read(*cmd, stderr=subprocess.PIPE)
+        except subprocess.CalledProcessError as e:
+            # As far as I'm aware, this only happens in a repo without any
+            # commits or a corrupted repo.
+            raise NoTagError(f"`git describe` command failed: {e.stderr.strip()}")
+
+    def get_branch(self) -> Optional[str]:
+        try:
+            return self.read("symbolic-ref", "--short", "-q", "HEAD")
+        except subprocess.CalledProcessError:
+            return None
 
 
 def describe_git(*, project_dir: Union[str, Path], **kwargs: Any) -> VCSDescription:
@@ -29,36 +75,31 @@ def describe_git(*, project_dir: Union[str, Path], **kwargs: Any) -> VCSDescript
         kwargs, "tool.versioningit.vcs", ["match", "exclude", "default-tag"]
     )
     build_date = get_build_date()
-
+    repo = GitRepo(project_dir)
+    repo.ensure_is_repo()
     try:
-        runcmd(
-            "git",
-            "-C",
-            project_dir,
-            "rev-parse",
-            "--git-dir",
-            stdout=subprocess.DEVNULL,
-            stderr=subprocess.DEVNULL,
-        )
-    except FileNotFoundError:
-        raise NotVCSError("Git not installed; assuming this isn't a Git repository")
-    except subprocess.CalledProcessError:
-        raise NotVCSError(f"{project_dir} is not a Git repository")
-
-    def readgit(*args: str, **kwargs: Any) -> str:
-        return readcmd("git", "-C", project_dir, *args, **kwargs)
-
-    describe_cmd = ["describe", "--tags", "--long", "--dirty", "--always"]
-    for pat in match:
-        describe_cmd.append(f"--match={pat}")
-    for pat in exclude:
-        describe_cmd.append(f"--exclude={pat}")
-    try:
-        description = readgit(*describe_cmd, stderr=subprocess.PIPE)
-    except subprocess.CalledProcessError as e:
-        # As far as I'm aware, this only happens in a repo without any commits
-        # or a corrupted repo.
-        raise NoTagError(f"`git describe` command failed: {e.stderr.strip()}")
+        description = repo.describe(match, exclude)
+    except NoTagError as e:
+        if default_tag is not None:
+            log.error("%s", e)
+            log.info("Falling back to default tag %r", default_tag)
+            return VCSDescription(
+                tag=default_tag,
+                state="dirty",
+                branch=repo.get_branch(),
+                fields={
+                    "distance": 0,
+                    "rev": "0" * 7,
+                    "revision": "0" * 40,
+                    "author_date": min(build_date, DEFAULT_DATE),
+                    "committer_date": min(build_date, DEFAULT_DATE),
+                    "build_date": build_date,
+                    "vcs": "g",
+                    "vcs_name": "git",
+                },
+            )
+        else:
+            raise
     if description.endswith("-dirty"):
         dirty = True
         description = description[: -len("-dirty")]
@@ -82,7 +123,7 @@ def describe_git(*, project_dir: Union[str, Path], **kwargs: Any) -> VCSDescript
             default_tag,
         )
         tag = default_tag
-        distance = int(readgit("rev-list", "--count", "HEAD")) - 1
+        distance = int(repo.read("rev-list", "--count", "HEAD")) - 1
         rev = description
     else:
         raise NoTagError("`git describe` could not find a tag")
@@ -94,18 +135,13 @@ def describe_git(*, project_dir: Union[str, Path], **kwargs: Any) -> VCSDescript
         state = "dirty"
     else:
         state = "exact"
-    revision, author_ts, committer_ts = readgit(
+    revision, author_ts, committer_ts = repo.read(
         "--no-pager", "show", "-s", "--format=%H%n%at%n%ct"
     ).splitlines()
-    branch: Optional[str]
-    try:
-        branch = readgit("symbolic-ref", "--short", "-q", "HEAD")
-    except subprocess.CalledProcessError:
-        branch = None
     return VCSDescription(
         tag=tag,
         state=state,
-        branch=branch,
+        branch=repo.get_branch(),
         fields={
             "distance": distance,
             "rev": rev,